Sends "content-type" to seldon deployment in .../predictions
Adds a new requirement:
filetype==1.0.7 - package to infer file type and MIME type

Infers the MIME type using the lib filetype, and passes it to
seldon deployment using a field in the body:
{..., "meta":{"content-type": "here-the-mime-type"}}

This is useful for tasks that receive a binary body (eg. images, audio),
but also need some metadata about the content (the MIME type).