Creates a single LogController for both experiments and deployments
By having a single controller (and a single response format) we make
it easier for the frontend to use the logs API.

This version returns only logs from containers (the previous version
also made a request to Jupyter to extract logs from Experiment.ipynb).
We'll change the approach and redirect notebook errors to the container,
and make this code significantly shorter and easier to understand.