This commit handles the direct translation of status_message. For this task I have two approaches(this commit implemented the approach (i)):
Approach (i): Directly translate in the message using a translation library, is the simplest way and easy to implement, the bad thing is that could be translation mistakes

Approach (ii): Use a more robust translation library, one that maps the message in multiple languages that way we won't have translation mistakes because the translation would be pre-defined in this map(locale data). The bad thing is that we would have to know the possible messages and that would take more time to implement and test. In addition, I would have to have a better understanding of the project to implement this solution. The good thing about this solution is that the project would be internationalized