Watching log stream and pod creation
Whenever the event source is called, there's a new thread for each pod that listen for new logs and there's a thread that watches for new pods being created. But there's a limitation within the log generation. When the client disconnects from the event source, the allocated threads aren't deallocated, not releasing the memory and process used.