Moves models/deployment.py SQLAlchemy ORM event to DeploymentController
models/deployment.py models used to import a RunController, but this
seems to be a bad pattern, since models should be only an 'Entity' and not
import/trigger much logic.
This also would cause recursive import problems if controller and kfp tried to import
models.

This update removes all controller imports from models/deployment.py, and moves those functions
to the Controller itself.