Feature/error codes (#358)
* Adds instructions on how to create/document error messages for APIs

Modifies CODE-REVIEW.md with these instructions.

* Adds Error Responses to docs/swagger.yaml

All errors have a code, a message, and a status (code).
This will improve the maintainability of our APIs when integrating
with other systems (eg. Frontend).

* Adds code and status_code to existing ApiExceptions

Also adds docstrings.

* Adds CannotCreateNamespacedPod to docs/swagger.yaml

* Adds code and status_code to InternalServerError, ServiceUnavailable, and Forbidden exceptions

* Fix mispell 'comparision' -> 'comparison'

* Adds code and status_code to NotFound exceptions

* Adds DeploymentNameExists response to swagger docs

* Adds code and status_code to BadRequest exceptions

* Replaces inadequate 'NotFound' exceptions by 'BadRequest' exceptions

HTTP Status 404 NotFound should be returned when a resource ID (provided in the URL)
does not exist.
When dealing with resources that are sent in the request body, one should
return HTTP Status 400 BadRequest.

* Adds DatasetNotFound exception to docs/swagger.yaml

* Adapts unittests to new returns 'code' and 'status_code'

* Removes unnecessary experiment.name validation

name is already validated by pydantic schema.

* Renames occurrences of 'status' to 'status_code' at CODE-REVIEW.md

* Replaces occurrences of '...NotFound' exceptions by 'Invalid...' exception

When a value is sent in request body, validation errors should be named 'Invalid...'
instead of '...NotFound'.

* Reorganizes docs/swagger.yaml

Adds examples to all endpoints.

* Adds unittests to test_deployments

Tests exception codes and creation using a templateId.

* Adds unittests to test_metrics

Tests return when there is a FileNotFoundError.

* Adds unittests to test_tasks

Covers validation errors what weren't tested.

* Removes status_code from responses body

After some discussion we decided that this is a redudant information
(what could be retrived using headers), so we removed from the response body.